Update mod.rs: make mod instr public. NOT FOR PROPOGATION without further testing and review.
There are use cases for user access to the instructions module, so not sure why it isn't public. It's certainly possible to screw up the operation of a running SM by using them, maybe it's a design goal to minimize these possibilities? Nevertheless it prevents useful and safe interactions between the CPU and PIO programs - I'd argue for a more general implementation, or at least more complete.